Added g band image inspections for CNN, NN, RF
All 3 prediction sets used the full training data of about 200k galaxies. I got the Δz of the photometric redshift predictions for the CNN with respect to the NN and RF.  Currently, the galaxies with the top 12 Δz's are shown in the g band. The next iteration will contain the galaxies with the smallest Δz's.